requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-autoslug==1.9.8
django-environ==0.6.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.18.2
factory-boy==3.2.0
Faker==8.12.1
flake8==3.9.2
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
